코드 리펙토링 진행 Warning 원인 파악 및 픽스(iterator 관련 제외) -> 픽스 Warning List[ C4100, C4305, C4189, C4018, C4239, C4244, C4245, C4305 ]

+++ Frameworks/Frameworks/Code/05.Objects/03.AnimatedObject/AnimatedObject.h
@@ -52,7 +52,7 @@ public: // 공개 함수
 
 	void  SetSpeed(float speed) { m_speed = speed; }
 
-	int GetAnimTimeRemain() { return m_nAniLength[m_nCurrAnimation] - m_fFrameTime; }
+	int GetAnimTimeRemain() { return static_cast<int>(m_nAniLength[m_nCurrAnimation] - m_fFrameTime); }
 	float GetAnimTimeRemainRatio() { return (m_nAniLength[m_nCurrAnimation] - m_fFrameTime) / (float)m_nAniLength[m_nCurrAnimation]; }
 
 protected: // 내부 함수
